Add tests for dealing with async callbacks
This adds a test to ensure that callbacks are called in the correct
order, and it alters the way that `didInsertNode` is called on child
elements. Previously they'd all be fired off at the same time, and would
race to finish, and `Promise.all` would ensure that the all finished
before proceeding. With this change we `await` each individual child to
ensure that they get called in the right order.

I'm not entirely sure if the children being called in the right order
is important. If it's not, then we could go back to using `Promise.all`
and could adjust the test accordingly.

(For my purposes the children being called in order is not important.
It's only important that the child callbacks are called _after_ the
parent callback has completed.)